content-automation/pyproject.toml
[project]
name = "content-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "anthropic>=0.89.0",
    "duckduckgo-search>=6.0.0",
    "fastapi>=0.135.3",
    "loguru>=0.7.3",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.24",
    "python-telegram-bot>=21.0",
    "uvicorn[standard]>=0.43.0",
]
